Update discord.js to version 14 and convert to using ES modules
What: Updates the discord.js dependency and all our other dependencies to their latest versions and converts the bot from a CommonJS module to an ES module.  All dependencies are loaded via the ES module system.  Node is updated to version 22.14.0.

Why: The version of discord.js, and all our other dependencies, was very old and at least discord.js will likely stop working in the near future.  The project was also pretty much impossible to build/run in this outdated state.  During the update, at least one of our dependencies wouldn't load using CommonJS and since ES modules is the recommended way to work with javascript libraries now, it seemed the best solution to convert to the ES modules system.

Notes:
Sharding hasn't been updated, because it would require fairly significant changes and the bot isn't used enough that we really need to utilize it yet.  Everything related to sharding has been commented out and left as a reference/todo for the future.

The `disabledEvents` parameter for the client constructor was removed in early 2020 in favor of intents.  Instead of opting out of events to improve performance you opt in to only the events you need (via intents) which should make the bot as performant as possible.
https://github.com/discordjs/discord.js/commit/8a2f8938be35e70f87ad99b3860116526e8622e6
https://discordjs.guide/popular-topics/intents.html

Statsd and Sentry haven't been tested, so it's possible they will need further updates in the future.