Update/Improve Readme
What: Updates the readme with missing steps and info and reformats a few things to be clearer.

Why: Without these changes it is very hard to figure out everything that is needed to set up and make the bot work fully.  Also, some of its capabilities and limitations were unclear.

Note: The section on deploying to your server likely needs updating as well, but isn't addressed in this commit.