Show Discord usernames on Hubs room join/leave
What: Shows the username in addition to the display name when a user joins or leaves the Hubs room if the Hubs room is bound to Discord.

Why: Since Discord display/nick names aren't guaranteed to be unique, usernames are the only reliable way to identify Discord users in the Hubs room.  Showing the username on join/leave makes it easier to identify users on the Discord side in case any moderation actions need to be taken, without being too intrusive.