Prevent the previously allowed users from killing the bot
What: Comments out the users that are allowed to kill the bot and effectively disables the kill command for the bot.

Why: It is no longer appropriate for the specified users to have the power to kill the Hubs Discord bot, especially given that they would have this power by default for all instances of the bot.  Since it hasn't been decided what will be done with this command, the users are just commented out for now.