Update RETICULUM_HOST default value in `.env`
What: Updates the value of the default RETICULUM_HOST for `.env` with a more appropriate value for setup against Community Edition instances.

Why: dev.reticulum.io is no longer available and a generic placeholder makes it clearer that it should be replaced with the domain of your instance while still conveying how it is expected to be formatted.