Correct the version of Node listed in the engines property
What: Changes the version of Node in the "engines" property of package.json to 22.14.0 to match what is listed in the readme.

Why: 22.14.0 is the actual version that was used when developing https://github.com/Hubs-Foundation/hubs-discord-bot/pull/145 and the current recommended version to use.  Review suggestion from https://github.com/Hubs-Foundation/hubs-discord-bot/pull/146#discussion_r2072677162